--- lib/auth/const.js
@@ -1,10 +1,0 @@
-'use strict';
-
-Object.defineProperty(exports, '__esModule', {
-	value: true
-});
-exports.TOKEN_TYPE_ACCESS = exports.TOKEN_TYPE_CHALLENGE = void 0;
-const TOKEN_TYPE_CHALLENGE = 'CHALLENGE';
-exports.TOKEN_TYPE_CHALLENGE = TOKEN_TYPE_CHALLENGE;
-const TOKEN_TYPE_ACCESS = 'ACCESS';
-exports.TOKEN_TYPE_ACCESS = TOKEN_TYPE_ACCESS;
