+++ lib/auth/index.js
@@ -3,43 +3,21 @@
 Object.defineProperty(exports, '__esModule', {
 	value: true
 });
-Object.defineProperty(exports, 'generateChallengeToken', {
-	enumerable: true,
-	get: function() {
-		return _generateChallengeToken.generateChallengeToken;
-	}
-});
-Object.defineProperty(exports, 'generateNonce', {
-	enumerable: true,
-	get: function() {
-		return _generateNonce.generateNonce;
-	}
-});
-Object.defineProperty(exports, 'generateAccessToken', {
-	enumerable: true,
-	get: function() {
-		return _generateAccessToken.generateAccessToken;
-	}
-});
-Object.defineProperty(exports, 'validateChallengeToken', {
-	enumerable: true,
-	get: function() {
-		return _validateChallengeToken.validateChallengeToken;
-	}
-});
-Object.defineProperty(exports, 'validateAccessToken', {
-	enumerable: true,
-	get: function() {
-		return _validateAccessToken.validateAccessToken;
-	}
-});
+exports.generateNonce = void 0;
 
-var _generateChallengeToken = require('./generate-challenge-token');
+var _crypto = _interopRequireDefault(require('crypto'));
 
-var _generateNonce = require('./generate-nonce');
+var _util = _interopRequireDefault(require('util'));
 
-var _generateAccessToken = require('./generate-access-token');
+function _interopRequireDefault(obj) {
+	return obj && obj.__esModule ? obj : { default: obj };
+}
 
-var _validateChallengeToken = require('./validate-challenge-token');
+const randomBytes = _util.default.promisify(_crypto.default.randomBytes);
 
-var _validateAccessToken = require('./validate-access-token');
+const generateNonce = async (length = 64) => {
+	const buffer = await randomBytes(length);
+	return buffer.toString('base64');
+};
+
+exports.generateNonce = generateNonce;
