+++ lib/index.js
@@ -16,18 +16,6 @@ Object.keys(_auth).forEach(function(key) {
 	});
 });
 
-var _jwt = require('./jwt');
-
-Object.keys(_jwt).forEach(function(key) {
-	if (key === 'default' || key === '__esModule') return;
-	Object.defineProperty(exports, key, {
-		enumerable: true,
-		get: function() {
-			return _jwt[key];
-		}
-	});
-});
-
 var _did = require('./did');
 
 Object.keys(_did).forEach(function(key) {
