--- lib/jwt/const.js
@@ -1,11 +1,0 @@
-'use strict';
-
-Object.defineProperty(exports, '__esModule', {
-	value: true
-});
-exports.JWT_ALGORITHMS = void 0;
-const JWT_ALGORITHMS = {
-	RSA: 'RS512',
-	HMAC: 'HS512'
-};
-exports.JWT_ALGORITHMS = JWT_ALGORITHMS;
