--- lib/jwt/index.js
@@ -1,29 +1,0 @@
-'use strict';
-
-Object.defineProperty(exports, '__esModule', {
-	value: true
-});
-Object.defineProperty(exports, 'parseJWT', {
-	enumerable: true,
-	get: function() {
-		return _parse.parseJWT;
-	}
-});
-Object.defineProperty(exports, 'validateJWT', {
-	enumerable: true,
-	get: function() {
-		return _validate.validateJWT;
-	}
-});
-Object.defineProperty(exports, 'issueJWT', {
-	enumerable: true,
-	get: function() {
-		return _issue.issueJWT;
-	}
-});
-
-var _parse = require('./parse');
-
-var _validate = require('./validate');
-
-var _issue = require('./issue');
