--- src/auth/const.js
@@ -1,2 +1,0 @@
-export const TOKEN_TYPE_CHALLENGE = 'CHALLENGE';
-export const TOKEN_TYPE_ACCESS = 'ACCESS';
