--- src/auth/generate-nonce.js
@@ -1,5 +1,0 @@
-import { generateHMACKey } from '../key';
-
-export const generateNonce = async length => {
-	return generateHMACKey(length, 'base64');
-};
