+++ src/auth/index.spec.js
@@ -1,0 +1,13 @@
+import { generateNonce } from './index';
+
+describe('Generate nonce', () => {
+	it('correct length', async () => {
+		expect((await generateNonce()).length).toBe(88);
+	});
+
+	it('generates different values', async () => {
+		const nonce1 = await generateNonce();
+		const nonce2 = await generateNonce();
+		expect(nonce1 !== nonce2).toBe(true);
+	});
+});
