+++ src/index.js
@@ -1,3 +1,2 @@
 export * from './auth';
-export * from './jwt';
 export * from './did';
