--- src/jwt/__fixtures__/private.pem
@@ -1,51 +1,0 @@
------BEGIN RSA PRIVATE KEY-----
-MIIJKQIBAAKCAgEAyJwk9jX34Nn9dzLyR+88Y22259E1c6Pw7cgPpja6GVYADXrN
-CwFSseV+cenA2/L39xgBqNtj4WKx+zeeRB9jwqJWI+ddAOs4UUPuEyL7XfZ7Msf6
-OT/7j2JKo9eu4mQ9bFzEOKpsq5R09vVeESZHz3a8LUA73qDtxy9bVqMAbg5zqMMr
-964Pijeu0UcqM6hvf41SqLkL52Ijpk2/Iir688K12AfZW7/wCR6VimCuFYo0ZbWD
-i5TYGzIK9CX9KFS48/jaCZSKZN1icAg/8z3/MSfRvKfxGNKGDayxO9H5Xl0w7adO
-XZ0YZfmVL+ok1deA0flUVTfnJT+rWherAdWhqYOK4Dor900yNpZAiNDYT2q7Kbb4
-jM6hfQAj2iTRus3mYLbZZsjjSXdrZGbV28YLQPkGmJSUae0MLyZbUNCrXfonMls/
-EdqcsUnzW7O9uPGA5WLTaB4kkysJUii6CoXpUVxxGjaofDoSavWBtQx+usy3NW3i
-lUc2ApIITB+UTUo7XRyfDxRXJPBJ8u3Sb3HrdZYw4hg3qqHLrI3rC1a4klJhhQQE
-S/+yYzA5dFLZGWtZQdajAzvcx0TmXb4LumX1d015IdyKB/IoPJyJqKZ9t2riom0F
-um4JU/A4nRlNFsQrs7iklIQpnKnShJhHjcTSgzbvGxlrszTMprKeOlxt4X8CAwEA
-AQKCAgEArE+SX/kTJQuY7x3PhYI0OKMwjAXjB6U1eMhyyK4RQ1nfSbTFEkLcqPka
-lvXVNheIbbEtG8D2Fc4UWlf3lUPxCGkv3D8KX5OrGAJlKALQLnFceM3UXABdL1ea
-DQlhqgu3xrAXyyR3mxk1Ljqrxcc/vy+ZirS1uBA6rPA8YDo6UE4Rgk/CIEheCLaz
-bgbWdPCkdMcKzOnFE8X0IRB+r/aC/4WrdBep12kR4INopydL4CyJvSMgDnCW/yUI
-y5Ul0dKT2ttRTRFNnj2aI+iyJViHLc2qFYWbTVrm0jAXO+6rZyZctIUUhsw8/LpY
-LoPLh/JHMBQdH+WYoSIDKKtTs4uwAshqYXEtiXEHuyVmabPJZ+7FbNrOO/gY7a4W
-EsTZwCnQ77lHCbuE5LWoyV50aeF/n7DU7UOldl5R/HKUUlfHbXfKy/EB4yQHPwW3
-A+njtQa+eCxeekakGjKxXQC1pj1d0tMMFGcPsPAWeCvmhp3ur0SFb3O2i9kX97DM
-zjGI7yIl0xAfZ/EslSS8HoRW9QwII4vbWSiFe612IAacvQyU3Zf1gvxbg33cab88
-9SHka0m8iGAmQoxCf0CJckQGJ2r1zq4ayrz6SZ56+yKYRm5SIxzMN2x3In8bm88A
-jzsj7hs2Vj5ZboaAu4WKvQp7RBQz5lWpTTx0k7qtJLQUJOUhOEECggEBAPFS9EiN
-9pcPYGkFcOd3k4QfA1ZtYvyD7K3pERyTXg2G5MoqiLPASQHVfihvZWjV31hiw8Dv
-8OqAa5/nA2kjzMeD2xaEjk1M0auHM7v4/8+RJ+OBSElYsuwhvz+5nNRvl4SLGQUb
-AjArWc6+ZOnN/cwQ4nkzGRXg7eQOfdxld61lV2vHHj0pZJwsTSWFSurDGB8yyNvR
-P+71ynDWod6lL+Gztx0nwvIcMH0iXQBjThFslLHeoqKc0F/ZON/5NfRbDqXfJru9
-W7M+0683eSpn6q4d6Q3ZUZqqg8+83kPBjQ7+0HsFmjqCjosxXpVNhHaFMoNy+1WG
-ZHaq04yNKelt2NkCggEBANTPVX08mNR5HXYmof0L1qilTTasIXqPDGfSxBt8t0oS
-o41nrEhVAecpmG07zaKZjZMVwE4JUEy+r4M/zKF+67y960CEpmpIo0snVSsDiojc
-2EeSasEaYkeMF+mNtBisBdc/oi/E/sdFrS6N90oMq2EThh8Y9S6II5MNEwy0jWrT
-MnfEfpTC8TORDccKGmHOcxgBQlkOjdR2UeShSJeXfq/Szy9SjvDPJ8lhQOlh9f/A
-06aMAoEId2GVMuwdIAsXkppsszh/Xji+CRck/qYMhBHZpID14MQZQY2gexe+1Jp4
-GaomL7btrlpEXQPnLdE9bE7IKrnv5NIaCrwClHRa1hcCggEBAODh/OLf6Ds9CQrV
-FzGcLMuiWunfZa8LATPPTTLj4XGZkvsSR/G57k7XNIlZxZa8NzWWU/Uno8njgmew
-T21a/ymtpiyFulYuj/TYOU2h7EDxc8p+X256ScHicOZhgdQ1aW0KcnZfhNp+i2It
-6rw2SZww39sx0t6gjQ9IseEAbVYYai71jyhPuoIOhpXAcN/0jjYEjUjK5lk4kz4z
-xQRFyyj44Fs7Al+8BTOXIjOKVuXF2YsnfUcr+sJXGrkc8zqS4EOSf/XTr/CJS5qs
-hy0NSZSYKuXM7k7DN7IOvtFFPL4MlxsNTXvYzhzIFOvtINORrFDyhWRdqKEs75a6
-taPhyYECggEAVkLYr1QRywZxB8RvIwhaca/XlZK+W1SCvc70SQQLs83MV/QvJH64
-r7/vlk+IciF1/qnxcnBaCVG+CSCiTNAKtE3zfwqF/k0FS+BcqHvd+s4yKQjTlwSf
-08qYosYau2l3UKHtlxluAM5gfZ90HqL2EPp+Ajr2n271pJTHJP5G2IdVTLDsclZ2
-uOVSRVz2BrFNe+x+vWSl3vo7BcbOiwwQgSqdatw3Lqt5wCpYyoE32vyJjlfwZ+dL
-pNV6d7TymSxr84aeOwd95Ng7DROk2R/3pNio0tFX04U9SjBD1QM4YIjQNmUWMDs1
-0kclwQBDVS5lqaV/lDEGepetxDnzfRqY7wKCAQAzNUCyU8BK/aoinNKAxGAN+zbd
-gm1LjBUTozVN5zr/u7+bfZOa48MxzVHRbBnT0YgJOlleK9c5WvAPfdseBM6r7THh
-NlXOZ5l7hKof0OAMyWG+afCmQT5n0JaxpD9B6T0a8UcGpM+EEHBWCW/rjE6RY1Ac
-rSfb5jvjdrpL4r143vox2Co4VmJc9hl1XH6II8RKD71yuwEWlXlXxmr8rAvnJeVT
-GQEH72rFJ7wFgtC3y+zMnmtXGQBb6bI23jJpzqB6YcMeH8dKB3Y65lMD4aBSLsY/
-IGy6eVeS3QFfdR6QsSXI6c5d+DcqdJG8VlNZXFUHAXS5ffiGSi5ThJB5ecWP
------END RSA PRIVATE KEY-----
