--- src/jwt/__fixtures__/public.pem
@@ -1,14 +1,0 @@
------BEGIN PUBLIC KEY-----
-MIICIjANBgkqhkiG9w0BAQEFAAOCAg8AMIICCgKCAgEAyJwk9jX34Nn9dzLyR+88
-Y22259E1c6Pw7cgPpja6GVYADXrNCwFSseV+cenA2/L39xgBqNtj4WKx+zeeRB9j
-wqJWI+ddAOs4UUPuEyL7XfZ7Msf6OT/7j2JKo9eu4mQ9bFzEOKpsq5R09vVeESZH
-z3a8LUA73qDtxy9bVqMAbg5zqMMr964Pijeu0UcqM6hvf41SqLkL52Ijpk2/Iir6
-88K12AfZW7/wCR6VimCuFYo0ZbWDi5TYGzIK9CX9KFS48/jaCZSKZN1icAg/8z3/
-MSfRvKfxGNKGDayxO9H5Xl0w7adOXZ0YZfmVL+ok1deA0flUVTfnJT+rWherAdWh
-qYOK4Dor900yNpZAiNDYT2q7Kbb4jM6hfQAj2iTRus3mYLbZZsjjSXdrZGbV28YL
-QPkGmJSUae0MLyZbUNCrXfonMls/EdqcsUnzW7O9uPGA5WLTaB4kkysJUii6CoXp
-UVxxGjaofDoSavWBtQx+usy3NW3ilUc2ApIITB+UTUo7XRyfDxRXJPBJ8u3Sb3Hr
-dZYw4hg3qqHLrI3rC1a4klJhhQQES/+yYzA5dFLZGWtZQdajAzvcx0TmXb4LumX1
-d015IdyKB/IoPJyJqKZ9t2riom0Fum4JU/A4nRlNFsQrs7iklIQpnKnShJhHjcTS
-gzbvGxlrszTMprKeOlxt4X8CAwEAAQ==
------END PUBLIC KEY-----
