--- src/jwt/__fixtures__/wrong_private.pem
@@ -1,27 +1,0 @@
------BEGIN RSA PRIVATE KEY-----
-MIIEnwIBAAKCAQA1har4ZePgy3LNdNV3hafOAYLWQX8gbCwXtDzNz4+ENu/BlEed
-5GpnrvNdGu04gZRaOntFwmYX6PJSsFrhPdATgmky91R7kRqf6ky06U66qRy5X8Jh
-EAqjyjOghWTY0WzpOMWrI4zha2bdQHgTrGgXDVRy2k0LnWbS8O3QMICHO9Hx+rom
-/1YnLvNALjo4tNnjEuMIvIjdFgVVfrAMY46U5RTsqaF5xr/UdWt1RzOJPLFdnBim
-qVO/0eQLnpO6lCna3ecntNaA+18MHuwt0lZCrVhTBxnKN9N/VyCzu877DqKApQIB
-ZulZMy2DgIBOJy2Dq7FdqtsJTDTLQcnqXhh5AgMBAAECggEAK7EezQCAcnYjFc4W
-pSeawPVC1Z1KJjGoHQGp0Wlrg0dFmgVuIVrh8hpD2PGuTb6aWL38/AmiRkqH5zDI
-P6+yOPrw3gZCFU1zKrCeReFfv8lmBX5Cs/Hgfa0HUDcVCFqOSiVqWDcBPtPqeRuz
-6nuHzam4ErpawuyPkAvWV2oAwmQ30q29tOR2nJRi2rA6pwROi7kX6klwicmHgG4H
-MEh/UtuUnsWqjWkAF9lA4T/svj9p+XZNTVfNrf4TWnCEFcT+g5BAhdfIOztz+Hhr
-3F9TZZbwYXnMe/Clmq8fWckTjZbUYjpzHiQ5qtqjdJ2N8XvjGA59Vel+5OqT2QWn
-Dg0ScQKBgHmX2fddfhp11hnvRaJoQAciqNqIyvc1RYLi3UVTgcLuOdTDAQ6OYx7X
-XTcEeen248smQ5A+NU4o9hLQ+vEjToy3UyzOkE1JE6iWedoH9PdPMEsjeGIMoHWA
-hkWiWTXgbwsnwyevOeRvnogoRHVTBm3uIEoMa08ZX9kiB6gQ3vprAoGAcK88pXOm
-k/tAiybmdkjGyw6farrtXR6Rx++5NcAPo10x7GnEZbvip7PqSLed4wIWQIjX6qzp
-kP9xGrStevDN95e6jw/uF2V6u3mCXj7YuI+kYAqp9sexH1pNbfIDKJ+d/j1iJJ7Q
-+ZX5j6LGliJ9k1USpDgzBmxAOgidVd/BOasCgYBvt5tmcZoJSbcbQQIjw9EAPVIS
-55hEJcfmB/bxWVZh+WtdFoYED8L2dT2J7SzuUA7fehqdjup/E5dzllrZ72yeUVm0
-U122BKxxO0BdmO3zw+hFQSO8QomzLBEyLYX1wG2GPFRMQ+zW+88Wnpi3iz7fBejY
-4CxAErpDz8ZlHosgXQKBgD8J2XyZqUvJ2D7AYVc0uXlpUceMkTNx4Y4fTtaB/n2e
-qkoelU9wf4jf/qX1iq8vQLGTnG7dksCaFnUhNJe9VQw45pYYDFMsYy+z9zmwIN+H
-BCbycEjtR+pJQ13JcPjt6p4LSvc/nZqyyu0O29erreKZShdylSD3uCp/HT2oMlmH
-AoGAd3zstmCV38yaDHjXivQXosxYu0whc8bJvg6N3Uwh7Tqfef6kX8AO40a0k4Kz
-QDQVcv9X1gOwCn/NrEhlHgAlB9U6kXd+2i9eR8/VtT0CpuHGy8Nf7LMWLg7hv2Vq
-wOe/cMvHuNaeh9olSfMun461c8pOmfYzJFtiZFacKynXp5w=
------END RSA PRIVATE KEY-----
