--- src/jwt/__fixtures__/wrong_public.pem
@@ -1,9 +1,0 @@
------BEGIN PUBLIC KEY-----
-MIIBITANBgkqhkiG9w0BAQEFAAOCAQ4AMIIBCQKCAQA1har4ZePgy3LNdNV3hafO
-AYLWQX8gbCwXtDzNz4+ENu/BlEed5GpnrvNdGu04gZRaOntFwmYX6PJSsFrhPdAT
-gmky91R7kRqf6ky06U66qRy5X8JhEAqjyjOghWTY0WzpOMWrI4zha2bdQHgTrGgX
-DVRy2k0LnWbS8O3QMICHO9Hx+rom/1YnLvNALjo4tNnjEuMIvIjdFgVVfrAMY46U
-5RTsqaF5xr/UdWt1RzOJPLFdnBimqVO/0eQLnpO6lCna3ecntNaA+18MHuwt0lZC
-rVhTBxnKN9N/VyCzu877DqKApQIBZulZMy2DgIBOJy2Dq7FdqtsJTDTLQcnqXhh5
-AgMBAAE=
------END PUBLIC KEY-----
