--- src/jwt/const.js
@@ -1,4 +1,0 @@
-export const JWT_ALGORITHMS = {
-	RSA: 'RS512',
-	HMAC: 'HS512'
-};
