--- src/jwt/index.js
@@ -1,3 +1,0 @@
-export { parseJWT } from './parse';
-export { validateJWT } from './validate';
-export { issueJWT } from './issue';
