--- src/jwt/parse.js
@@ -1,7 +1,0 @@
-import jwt from 'jsonwebtoken';
-
-export const parseJWT = token => {
-	return jwt.decode(token, { complete: true, json: true });
-};
-
-export default parseJWT;
