--- src/key/index.js
@@ -1,3 +1,0 @@
-export { generateHMACKey } from './hmac';
-export { generateRSAKeyPair } from './rsa';
-export { checkSecretLength } from './validate-private-key-size';
