--- src/key/rsa.js
@@ -1,32 +1,0 @@
-const { generateKeyPair } = require('crypto');
-const pemtools = require('pemtools');
-
-export const generateRSAKeyPair = (length = 4096) => {
-	return new Promise((resolve, reject) => {
-		generateKeyPair(
-			'rsa',
-			{
-				modulusLength: length,
-				publicKeyEncoding: {
-					type: 'spki',
-					format: 'pem'
-				},
-				privateKeyEncoding: {
-					type: 'pkcs1',
-					format: 'pem'
-				}
-			},
-			(err, publicKey, privateKey) => {
-				if (err) {
-					return reject(err);
-				}
-				return resolve({ privateKey, publicKey });
-			}
-		);
-	});
-};
-
-export const getPEMBits = privateKey => {
-	const pem = pemtools(privateKey);
-	return pem.pubkey.bits;
-};
