--- test/utils.js
@@ -1,6 +1,0 @@
-import fs from 'fs';
-import path from 'path';
-
-export const readTextFile = filename => {
-	return fs.readFileSync(path.resolve(filename), 'utf8');
-};
